+++ vault/resource_kv_secret_v2_test.go
@@ -215,6 +215,58 @@ func TestAccKVSecretV2_DisableRead(t *testing.T) {
 	})
 }
 
+// Fadia u have added this
+func TestAccKVSecretV2_UpdateOutsideTerraform(t *testing.T) {
+	resourceName := "vault_kv_secret_v2.test"
+	mount := acctest.RandomWithPrefix("tf-kv")
+	name := acctest.RandomWithPrefix("foo")
+
+	resource.Test(t, resource.TestCase{
+		ProviderFactories: providerFactories,
+		PreCheck:          func() { testutil.TestAccPreCheck(t) },
+		Steps: []resource.TestStep{
+			{
+				Config: testKVSecretV2Config_initial(mount, name),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, consts.FieldMount, mount),
+					resource.TestCheckResourceAttr(resourceName, consts.FieldName, name),
+					resource.TestCheckResourceAttr(resourceName, consts.FieldPath, fmt.Sprintf("%s/data/%s", mount, name)),
+					resource.TestCheckResourceAttr(resourceName, "delete_all_versions", "true"),
+					resource.TestCheckResourceAttr(resourceName, "data.zip", "zap"),
+					resource.TestCheckResourceAttr(resourceName, "data.foo", "bar"),
+					resource.TestCheckResourceAttr(resourceName, "data.flag", "false"),
+					resource.TestCheckResourceAttr(resourceName, "metadata.version", "1"),
+				),
+			},
+			{
+				PreConfig: func() {
+					client := testProvider.Meta().(*provider.ProviderMeta).MustGetClient()
+
+					// Simulate external change using Vault CLI
+					path := fmt.Sprintf("%s/data/%s", mount, name)
+					_, err := client.Logical().Write(path, map[string]interface{}{"data": map[string]interface{}{"testkey3": "testvalue3"}})
+					if err != nil {
+						t.Fatalf(fmt.Sprintf("error simulating external change; err=%s", err))
+					}
+
+				},
+
+				Config: testKVSecretV2Config_initial(mount, name),
+
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "data.zip", "zap"),
+					resource.TestCheckResourceAttr(resourceName, "data.foo", "bar"),
+					resource.TestCheckResourceAttr(resourceName, "data.flag", "false"),
+					resource.TestCheckResourceAttr(resourceName, "data_json", "{\"flag\":false,\"foo\":\"bar\",\"zip\":\"zap\"}"),
+					//we check that the provider updated vault to match the the terraform config therefor creating a new version the secret.
+					resource.TestCheckResourceAttr(resourceName, "metadata.version", "3"),
+				),
+			},
+		},
+	},
+	)
+}
+
 func readKVData(t *testing.T, mount, name string) {
 	t.Helper()
 	client := testProvider.Meta().(*provider.ProviderMeta).MustGetClient()
@@ -235,6 +287,7 @@ func readKVData(t *testing.T, mount, name string) {
 	if !reflect.DeepEqual(resp.Data["data"], testKVV2Data) {
 		t.Fatalf("kvv2 secret data does not match got: %#+v, want: %#+v", resp.Data["data"], testKVV2Data)
 	}
+
 }
 
 func writeKVData(t *testing.T, mount, name string) {